Use the ToDo App to work with Redux
