+++ App.js
@@ -1,25 +1,11 @@
-import React from 'react';
-import { View, StyleSheet, Text } from 'react-native';
-import ComponentA from './components/ComponentA';
-import ComponentB from './components/ComponentB';
-import store from './store/store';
-import { Provider } from 'react-redux';
+import React from 'react'
+import { NavigationContainer } from '@react-navigation/native';
+import MyDrawer from './routes/MyDrawer';
 
-export default function App() {
-  return (
-    <View style={styles.container}>
-      <Provider store={store}>
-        <ComponentA />
-        <ComponentB />
-      </Provider>
-     </View>
+export default function App(){
+  return(
+    <NavigationContainer> 
+      <MyDrawer/>
+    </NavigationContainer>
   );
 }
-
-const styles = StyleSheet.create({
-  container:{
-    alignItems: 'center',
-    flex: 1,
-    justifyContent: 'center'
-  }
-})
